requirements.txt
discord.py>=2.4.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
google-genai>=0.4.0,<1.0.0
PyNaCl>=1.5.0,<2.0.0
aiohttp>=3.9.0,<4.0.0
kokoro-tts
torch
soundfile
sounddevice